requirements.txt
asgiref==3.8.1
build==1.2.1
click==8.1.7
Django==5.0.4
djangorestframework==3.15.1
packaging==24.0
pip-tools==7.4.1
pyproject_hooks==1.0.0
setuptools==69.5.1
sqlparse==0.5.0
wheel==0.43.0
